removed bundle zip file process
